+++ client/src/components/organisms/account/AccountSelectListItem/index.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { IAccount } from 'types/account';
+import { formatBankCode } from 'utils/common/formatBankCode';
+import { ReactComponent as Shinhan } from 'assets/icons/banklogo/088.svg';
+import { ReactComponent as Change } from 'assets/icons/account/change.svg';
+import { AccountSelectListItemContainer } from './style';
+
+function AccountSelectListItem({ account, onlyView }: { account: IAccount; onlyView?: boolean }) {
+	const formattedBankCode = formatBankCode(account.bankCode);
+	return (
+		<AccountSelectListItemContainer>
+			<div className="account">
+				<Shinhan />
+				{account.alias}{' '}
+				<span>
+					{formattedBankCode} {account.accountNumber}
+				</span>
+			</div>
+			{onlyView ? (
+				<div />
+			) : (
+				<div className="change">
+					<Change />
+				</div>
+			)}
+		</AccountSelectListItemContainer>
+	);
+}
+
+export default AccountSelectListItem;
