+++ client/src/components/organisms/account/AccountSelector/index.tsx
@@ -1,45 +1,8 @@
 import React, { Dispatch, SetStateAction } from 'react';
-import { ReactComponent as Shinhan } from 'assets/icons/banklogo/088.svg';
-import { ReactComponent as Change } from 'assets/icons/account/change.svg';
 import { IAccount } from 'types/account';
-import { formatBankCode } from 'utils/common/formatBankCode';
 import { AccountSelectorContainer } from './style';
-
-function AccountSelectListItem({ account }: { account: IAccount }) {
-	const formattedBankCode = formatBankCode(account.bankCode);
-	return (
-		<li>
-			<div className="account">
-				<Shinhan />
-				{account.alias}{' '}
-				<span>
-					{formattedBankCode} {account.accountNumber}
-				</span>
-			</div>
-			<div className="change">
-				<Change />
-			</div>
-		</li>
-	);
-}
-
-function AccountSelectList({
-	accounts,
-	setSelected,
-}: {
-	accounts: IAccount[];
-	setSelected: Dispatch<SetStateAction<IAccount>>;
-}) {
-	return (
-		<div>
-			{accounts.map((el) => (
-				<div key={el.key} onClick={() => setSelected(el)} role="presentation">
-					<AccountSelectListItem account={el} />
-				</div>
-			))}
-		</div>
-	);
-}
+import AccountSelectListItem from '../AccountSelectListItem';
+import AccountSelectList from '../AccountSelectList';
 
 interface IAccountSelectorProps {
 	accounts: IAccount[];
